+++ tests/assertions/assert_attribute.test.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const Wendigo = require('../../lib/wendigo');
+const Wendigo = require('../..');
 const utils = require('../test_utils');
 const configUrls = require('../config.json').urls;