+++ tests/browser/query.test.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const Wendigo = require('../../lib/wendigo');
+const Wendigo = require('../..');
 const assert = require('assert');
 const DomElement = require('../../lib/models/dom_element');
 const configUrls = require('../config.json').urls;